requirements.txt
Mezzanine==4.2.3
psycopg2
future
fabric
django-recaptcha==1.3.1
zeep==2.3.0
djangorestframework